restricting users complete

+++ base/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render,redirect # type: ignore
 from django.contrib import messages #type:ignore
 from django.http import HttpResponse         # type: ignore
+from django.contrib.auth.decorators import login_required #type:ignore
 from django.db.models import Q # type: ignore
 from django.contrib.auth.models import User # type: ignore
 from django.contrib.auth import authenticate,login,logout #type:ignore
@@ -10,6 +11,10 @@ from .forms import RoomForm
 
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method=='POST':
         username=request.POST.get('username')
         password=request.POST.get('password')
@@ -57,6 +62,7 @@ def room(request,pk):
     context={'room':room}
     return render(request,'base/room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form=RoomForm()
     if request.method=='POST':
@@ -67,9 +73,14 @@ def createRoom(request):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
+
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You dont have the authority to update this room")
 
     if request.method=='POST':
         form=RoomForm(request.POST,instance=room)
@@ -80,8 +91,14 @@ def updateRoom(request,pk):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You dont have the authority to delete this room")
+    
     if request.method=='POST':
         room.delete()
         return redirect('home')
